pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "liveapi"
version = "0.1.0"
description = "API lifecycle management with versioning, change detection, and implementation sync."
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "liveapi", from = "src"}]

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.115.0"
pydantic = "^2.11.0"
pyyaml = "^6.0"
uvicorn = "^0.20.0"
prance = "^25.0.0"
jinja2 = "^3.1.0"
requests = "^2.31.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
pytest-asyncio = "^0.21.0"
pytest-cov = "^4.1.0"
httpx = "^0.24.0"
flake8 = "^7.0.0"
twine = "^4.0.0"
black = "^23.0.0"

[tool.poetry.scripts]
liveapi = "liveapi.cli:main"

